refactor: update service interfaces and implementations for Attendance, Attire, Category, Measurement, and User
