+++ src/main/java/com/example/FabriqBackend/controller/AttireRentController.java
@@ -1,7 +1,7 @@
 package com.example.FabriqBackend.controller;
 
 import com.example.FabriqBackend.dto.AttireRentAddDto;
-import com.example.FabriqBackend.service.AttireRentService;
+import com.example.FabriqBackend.service.IAttireRentService;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -12,7 +12,7 @@ import org.springframework.web.bind.annotation.*;
 @RequiredArgsConstructor
 public class AttireRentController {
 
-    private final AttireRentService attireRentService;
+    private final IAttireRentService attireRentService;
 
     @PostMapping("/add")
     @Operation(
