+++ src/main/java/com/example/FabriqBackend/controller/CategoryController.java
@@ -1,7 +1,7 @@
 package com.example.FabriqBackend.controller;
 
 import com.example.FabriqBackend.model.Category;
-import com.example.FabriqBackend.service.CategoryService;
+import com.example.FabriqBackend.service.ICategoryService;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -12,7 +12,7 @@ import org.springframework.web.bind.annotation.*;
 @RequiredArgsConstructor
 public class CategoryController {
 
-    private final CategoryService categoryService;
+    private final ICategoryService categoryService;
 
     @PostMapping("/add")
     @Operation(
